+++ terminusdb_client/tests/integration_tests/test_client.py
@@ -12,10 +12,12 @@ from terminusdb_client.woqlquery.woql_query import WOQLQuery
 
 # from terminusdb_client.woqlquery.woql_query import WOQLQuery
 
+test_user_agent = "terminusdb-client-python-tests"
+
 
 def test_happy_path(docker_url):
     # create client
-    client = WOQLClient(docker_url)
+    client = WOQLClient(docker_url, user_agent=test_user_agent)
     assert not client._connected
     # test connect
     client.connect()
@@ -60,7 +62,7 @@ def test_happy_path(docker_url):
 
 def test_happy_carzy_path(docker_url):
     # create client
-    client = WOQLClient(docker_url)
+    client = WOQLClient(docker_url, user_agent=test_user_agent)
     assert not client._connected
     # test connect
     client.connect()
@@ -93,11 +95,15 @@ def test_happy_carzy_path(docker_url):
 
 def test_diff_ops(docker_url, test_schema):
     # create client and db
-    client = WOQLClient(docker_url)
+    client = WOQLClient(docker_url, user_agent=test_user_agent)
     client.connect()
     client.create_database("test_diff_ops")
-    public_diff = WOQLClient("https://cloud.terminusdb.com/jsondiff")
-    public_patch = WOQLClient("https://cloud.terminusdb.com/jsonpatch")
+    public_diff = WOQLClient(
+        "https://cloud.terminusdb.com/jsondiff", user_agent=test_user_agent
+    )
+    public_patch = WOQLClient(
+        "https://cloud.terminusdb.com/jsonpatch", user_agent=test_user_agent
+    )
 
     result_patch = Patch(
         json='{"@id": "Person/Jane", "name" : { "@op" : "SwapValue", "@before" : "Jane", "@after": "Janine" }}'
@@ -138,7 +144,9 @@ def test_diff_ops(docker_url, test_schema):
     new_data_version = client.get_document(jane_id, get_data_version=True)[-1]
     new_commit = client._get_current_commit()
     commit_id_result2 = client.diff(current_commit, new_commit, document_id=jane_id)
-    data_version_result2 = client.diff(data_version, new_data_version, document_id=jane_id)
+    data_version_result2 = client.diff(
+        data_version, new_data_version, document_id=jane_id
+    )
     # test all diff commit_id and data_version
     commit_id_result_all = client.diff(current_commit, new_commit)
     data_version_result_all = client.diff(data_version, new_data_version)
@@ -206,7 +214,9 @@ def test_diff_ops(docker_url, test_schema):
 )
 def test_diff_ops_no_auth(test_schema, terminusx_token):
     # create client and db
-    client = WOQLClient("https://cloud-dev.terminusdb.com/TerminusDBTest//")
+    client = WOQLClient(
+        "https://cloud-dev.terminusdb.com/TerminusDBTest//", user_agent=test_user_agent
+    )
     client.connect(use_token=True, team="TerminusDBTest")
 
     result_patch = Patch(
@@ -252,7 +262,7 @@ def test_jwt(docker_url_jwt):
     # create client
     url = docker_url_jwt[0]
     token = docker_url_jwt[1]
-    client = WOQLClient(url)
+    client = WOQLClient(url, user_agent=test_user_agent)
     assert not client._connected
     # test connect
     client.connect(use_token=True, jwt_token=token)
@@ -274,7 +284,7 @@ def test_terminusx(terminusx_token):
         "test_happy_" + str(dt.datetime.now()).replace(" ", "") + "_" + str(random())
     )
     endpoint = "https://cloud-dev.terminusdb.com/TerminusDBTest/"
-    client = WOQLClient(endpoint)
+    client = WOQLClient(endpoint, user_agent=test_user_agent)
     client.connect(use_token=True, team="TerminusDBTest")
     assert client._connected
     # test create db
@@ -292,7 +302,7 @@ def test_terminusx(terminusx_token):
 def test_terminusx_crazy_path(terminusx_token):
     testdb = "test happy_" + str(dt.datetime.now()).replace(" ", "")
     endpoint = "https://cloud-dev.terminusdb.com/TerminusDBTest/"
-    client = WOQLClient(endpoint)
+    client = WOQLClient(endpoint, user_agent=test_user_agent)
     client.connect(use_token=True, team="TerminusDBTest")
     assert client._connected
     # test create db