+++ terminusdb_client/woqlclient/woqlClient.py
@@ -154,13 +154,20 @@ class WOQLClient:
         Repo identifier of the database that this client is connected to. Default to "local".
     """
 
-    def __init__(self, server_url: str, **kwargs) -> None:
+    def __init__(
+        self,
+        server_url: str,
+        user_agent: str = f"terminusdb-client-python/{__version__}",
+        **kwargs,
+    ) -> None:
         r"""The WOQLClient constructor.
 
         Parameters
         ----------
         server_url : str
             URL of the server that this client will connect to.
+        user_agent: optional, str
+            User agent header when making requests. Defaults to terminusdb-client-python with the version appended.
         \**kwargs
             Extra configuration options
 
@@ -176,6 +183,9 @@ class WOQLClient:
         self._branch = None
         self._ref = None
         self._repo = None
+
+        # Default headers
+        self._default_headers = {"user-agent": user_agent}
 
     @property
     def team(self):
@@ -331,9 +341,7 @@ class WOQLClient:
                 _finish_response(
                     requests.get(
                         self.api + "/info",
-                        headers={
-                            "user-agent": f"terminusdb-client-python/{__version__}"
-                        },
+                        headers=self._default_headers,
                         auth=self._auth(),
                     )
                 )
@@ -347,9 +355,7 @@ class WOQLClient:
                 _finish_response(
                     requests.head(
                         self._db_url(),
-                        headers={
-                            "user-agent": f"terminusdb-client-python/{__version__}"
-                        },
+                        headers=self._default_headers,
                         params={"exists": "true"},
                         auth=self._auth(),
                     )
@@ -485,7 +491,7 @@ class WOQLClient:
         api_url = "/".join(api_url) + "/_commits"
         result = requests.get(
             api_url,
-            headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+            headers=self._default_headers,
             params={"type": "Branch"},
             auth=self._auth(),
         )
@@ -609,7 +615,7 @@ class WOQLClient:
         self._check_connection()
         result = requests.get(
             self._db_base("prefixes"),
-            headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+            headers=self._default_headers,
             auth=self._auth(),
         )
         return json.loads(_finish_response(result))
@@ -682,7 +688,7 @@ class WOQLClient:
         _finish_response(
             requests.post(
                 self._db_url(),
-                headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+                headers=self._default_headers,
                 json=details,
                 auth=self._auth(),
             )
@@ -738,7 +744,7 @@ class WOQLClient:
         _finish_response(
             requests.delete(
                 self._db_url(),
-                headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+                headers=self._default_headers,
                 auth=self._auth(),
                 params=payload,
             )
@@ -774,7 +780,7 @@ class WOQLClient:
         self._validate_graph_type(graph_type)
         result = requests.get(
             self._triples_url(graph_type),
-            headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+            headers=self._default_headers,
             auth=self._auth(),
         )
         return json.loads(_finish_response(result))
@@ -807,7 +813,7 @@ class WOQLClient:
         params["turtle"] = turtle
         result = requests.post(
             self._triples_url(graph_type),
-            headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+            headers=self._default_headers,
             params=params,
             auth=self._auth(),
         )
@@ -843,7 +849,7 @@ class WOQLClient:
         params["turtle"] = turtle
         result = requests.put(
             self._triples_url(graph_type),
-            headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+            headers=self._default_headers,
             params=params,
             auth=self._auth(),
         )
@@ -892,13 +898,11 @@ class WOQLClient:
         for the_arg in add_args:
             if the_arg in kwargs:
                 payload[the_arg] = kwargs[the_arg]
-
+        headers = self._default_headers.copy()
+        headers["X-HTTP-Method-Override"] = "GET"
         result = requests.post(
             self._documents_url(),
-            headers={
-                "user-agent": f"terminusdb-client-python/{__version__}",
-                "X-HTTP-Method-Override": "GET",
-            },
+            headers=headers,
             json=payload,
             auth=self._auth(),
         )
@@ -956,7 +960,7 @@ class WOQLClient:
 
         result = requests.get(
             self._documents_url(),
-            headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+            headers=self._default_headers,
             params=payload,
             auth=self._auth(),
         )
@@ -1019,7 +1023,7 @@ class WOQLClient:
                 payload[the_arg] = kwargs[the_arg]
         result = requests.get(
             self._documents_url(),
-            headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+            headers=self._default_headers,
             params=payload,
             auth=self._auth(),
         )
@@ -1087,7 +1091,7 @@ class WOQLClient:
                 payload[the_arg] = kwargs[the_arg]
         result = requests.get(
             self._documents_url(),
-            headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+            headers=self._default_headers,
             params=payload,
             auth=self._auth(),
         )
@@ -1225,7 +1229,7 @@ class WOQLClient:
         else:
             params["full_replace"] = "false"
 
-        headers = {"user-agent": f"terminusdb-client-python/{__version__}"}
+        headers = self._default_headers.copy()
         if last_data_version is not None:
             headers["TerminusDB-Data-Version"] = last_data_version
 
@@ -1318,7 +1322,7 @@ class WOQLClient:
         params["graph_type"] = graph_type
         params["create"] = "true" if create else "false"
 
-        headers = {"user-agent": f"terminusdb-client-python/{__version__}"}
+        headers = self._default_headers.copy()
         if last_data_version is not None:
             headers["TerminusDB-Data-Version"] = last_data_version
 
@@ -1433,7 +1437,7 @@ class WOQLClient:
         params = self._generate_commit(commit_msg)
         params["graph_type"] = graph_type
 
-        headers = {"user-agent": f"terminusdb-client-python/{__version__}"}
+        headers = self._default_headers.copy()
         if last_data_version is not None:
             headers["TerminusDB-Data-Version"] = last_data_version
 
@@ -1465,7 +1469,7 @@ class WOQLClient:
         self._validate_graph_type(graph_type)
         self._check_connection()
         all_existing_obj = self.get_all_documents(graph_type=graph_type)
-        all_existing_id = list(map(lambda x: x.get("@id"), all_existing_obj))
+        all_existing_id = [x.get("@id") for x in all_existing_obj]
         return doc_id in all_existing_id
 
     def get_class_frame(self, class_name):
@@ -1485,7 +1489,7 @@ class WOQLClient:
         opts = {"type": class_name}
         result = requests.get(
             self._class_frame_url(),
-            headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+            headers=self._default_headers,
             params=opts,
             auth=self._auth(),
         )
@@ -1538,7 +1542,7 @@ class WOQLClient:
             request_woql_query = woql_query
         query_obj["query"] = request_woql_query
 
-        headers = {"user-agent": f"terminusdb-client-python/{__version__}"}
+        headers = self._default_headers.copy()
         if last_data_version is not None:
             headers["TerminusDB-Data-Version"] = last_data_version
 
@@ -1589,7 +1593,7 @@ class WOQLClient:
         _finish_response(
             requests.post(
                 self._branch_url(new_branch_id),
-                headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+                headers=self._default_headers,
                 json=source,
                 auth=self._auth(),
             )
@@ -1613,7 +1617,7 @@ class WOQLClient:
         _finish_response(
             requests.delete(
                 self._branch_url(branch_id),
-                headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+                headers=self._default_headers,
                 auth=self._auth(),
             )
         )
@@ -1670,7 +1674,7 @@ class WOQLClient:
 
         result = requests.post(
             self._pull_url(),
-            headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+            headers=self._default_headers,
             json=rc_args,
             auth=self._auth(),
         )
@@ -1693,7 +1697,7 @@ class WOQLClient:
 
         result = requests.post(
             self._fetch_url(remote_id),
-            headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+            headers=self._default_headers,
             auth=self._auth(),
         )
 
@@ -1750,7 +1754,7 @@ class WOQLClient:
 
         result = requests.post(
             self._push_url(),
-            headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+            headers=self._default_headers,
             json=rc_args,
             auth=self._auth(),
         )
@@ -1817,7 +1821,7 @@ class WOQLClient:
 
         result = requests.post(
             self._rebase_url(),
-            headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+            headers=self._default_headers,
             json=rc_args,
             auth=self._auth(),
         )
@@ -1875,7 +1879,7 @@ class WOQLClient:
         _finish_response(
             requests.post(
                 self._reset_url(),
-                headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+                headers=self._default_headers,
                 json={"commit_descriptor": commit_path},
                 auth=self._auth(),
             )
@@ -1910,7 +1914,7 @@ class WOQLClient:
         _finish_response(
             requests.post(
                 self._optimize_url(path),
-                headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+                headers=self._default_headers,
                 auth=self._auth(),
             )
         )
@@ -1956,7 +1960,7 @@ class WOQLClient:
 
         result = requests.post(
             self._squash_url(),
-            headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+            headers=self._default_headers,
             json={"commit_info": self._generate_commit(message, author)},
             auth=self._auth(),
         )
@@ -2000,7 +2004,7 @@ class WOQLClient:
             "DocumentTemplate",  # noqa:F821
             List["DocumentTemplate"],  # noqa:F821
         ],
-        document_id: Union[str, None] = None
+        document_id: Union[str, None] = None,
     ):
         """Perform diff on 2 set of document(s), result in a Patch object.
 
@@ -2026,17 +2030,21 @@ class WOQLClient:
                 request_dict[key] = self._convert_diff_dcoument(item)
         if document_id is not None:
             if "before_data_version" in request_dict:
-                if document_id[:len("terminusdb:///data")] == "terminusdb:///data":
+                if document_id[: len("terminusdb:///data")] == "terminusdb:///data":
                     request_dict["document_id"] = document_id
                 else:
-                    raise ValueError(f"Valid document id starts with `terminusdb:///data`, but got {document_id}")
+                    raise ValueError(
+                        f"Valid document id starts with `terminusdb:///data`, but got {document_id}"
+                    )
             else:
-                raise ValueError("`document_id` can only be used in conjusction with a data version or commit ID as `before`, not a document object")
+                raise ValueError(
+                    "`document_id` can only be used in conjusction with a data version or commit ID as `before`, not a document object"
+                )
         if self._connected:
             result = _finish_response(
                 requests.post(
                     self._diff_url(),
-                    headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+                    headers=self._default_headers,
                     json=request_dict,
                     auth=self._auth(),
                 )
@@ -2045,7 +2053,7 @@ class WOQLClient:
             result = _finish_response(
                 requests.post(
                     self.server_url,
-                    headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+                    headers=self._default_headers,
                     json=request_dict,
                 )
             )
@@ -2089,7 +2097,7 @@ class WOQLClient:
             result = _finish_response(
                 requests.post(
                     self._patch_url(),
-                    headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+                    headers=self._default_headers,
                     json=request_dict,
                     auth=self._auth(),
                 )
@@ -2098,7 +2106,7 @@ class WOQLClient:
             result = _finish_response(
                 requests.post(
                     self.server_url,
-                    headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+                    headers=self._default_headers,
                     json=request_dict,
                 )
             )
@@ -2136,7 +2144,7 @@ class WOQLClient:
         _finish_response(
             requests.post(
                 self._clone_url(newid),
-                headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+                headers=self._default_headers,
                 json=rc_args,
                 auth=self._auth(),
             )
@@ -2227,7 +2235,7 @@ class WOQLClient:
 
         result = requests.get(
             self.api + "/",
-            headers={"user-agent": f"terminusdb-client-python/{__version__}"},
+            headers=self._default_headers,
             auth=self._auth(),
         )
         return json.loads(_finish_response(result))